[Refactor] Part property editor cleanup
